feat: add async batched state indexer test
photon-api update

photon-api update

cleanup

test_indexer: set prove_by_index to false to make it aligned with photon

fix: adjust loop iteration in batched state indexer test for correct batch size

add logs for batch updates

refactor: change merkle_tree type to String and update related code

refactor: update log messages for clarity

refactor: update logging in tests and adjust prover configuration

wip

switch to zkp batch size of 500, regenerate accounts

switch back to zkp batch size of 10, regenerate accounts

test is green

add assertions for the case when indexer root & onchain root are different

add randomized num of inputs

refactor: update payer handling in batched state async indexer tests

refactor: enhance indexer handling and logging in tests and instructions

refactor: enhance account counters in compress and transfer functions

feat: add v2 support for compressed token accounts and enhance account structures

fix get_compressed_token_accounts_v2 deserialization

refactor: improve assertions and error handling in batched state async indexer tests

refactor: add limit_ledger_size option to init-validator helper

regenerate photon-api

photon-api regenerated

test: add v1 address to photon e2e test (#1616)

* test: add v1 address to test_state_indexer_async_batched

* regenerate photon-api

* test: add v1 address to test_state_indexer_async_batched

* fix: correct function name and improve proof handling

* manual run for prover & photon

* photon-api regenerated

* format

---------


refactor: batch event into unified MerkleTreeEvent enum  (#1618)

* refactor: clean up batched state async indexer tests and improve timeout duration

* refactor: batched Merkle tree event into Merkle tree enum

* fix: batch Merkle tree simulate

* Update batched_state_test.rs

* Update batched_state_indexer_test.rs

* format

---------


account with test batch sizes

refactor: update AccountState to implement Display trait and adjust sleep durations in tests

refactor: update LightValidatorConfig to enable indexer and adjust prover configuration in tests

fix: implement get_indexer_slot to return a default value

refactor: update get_indexer_slot to accept RPC parameter and handle errors

format

add limit_ledger_size parameter to compressed-token-test configuration

refactor: adjust sleep durations in batched state async indexer tests

sdk-libs/tests: add limit_ledger_size field to RPC client configuration

refactor: remove debug logging from insert_into_queues

refactor: remove unnecessary reference to payer in create_and_send_transaction

fix: set proof by index correctly

test: enable indexer in batched state indexer tests

refactor: replace println! with log::debug! for improved logging consistency

refactor: replace log::debug! with tracing::debug! for improved logging consistency

refactor: enhance API request handling by appending API key to URIs

add forester e2e test to ci

cleanup

cleanup

+++ program-tests/utils/src/e2e_test_env.rs
@@ -764,7 +764,7 @@ where
                                     .get_queue_elements(
                                         merkle_tree_pubkey.to_bytes(),
                                         QueueType::BatchedAddress,
-                                        batch.batch_size,
+                                        batch.batch_size as u16,
                                         None,
                                     )
                                     .await
