+++ sdk-libs/photon-api/src/models/_get_queue_elements_post_request_params.rs
@@ -16,7 +16,7 @@ pub struct GetQueueElementsPostRequestParams {
     #[serde(rename = "merkleTree")]
     pub merkle_tree: String,
     #[serde(rename = "numElements")]
-    pub num_elements: u64,
+    pub num_elements: u16,
     #[serde(rename = "queueType")]
     pub queue_type: u16,
     #[serde(
@@ -30,7 +30,7 @@ pub struct GetQueueElementsPostRequestParams {
 impl GetQueueElementsPostRequestParams {
     pub fn new(
         merkle_tree: String,
-        num_elements: u64,
+        num_elements: u16,
         queue_type: u16,
     ) -> GetQueueElementsPostRequestParams {
         GetQueueElementsPostRequestParams {
